zporta_academy_backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Deprecated==1.2.18
Django==5.1.6
django-ckeditor-5==0.2.17
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
django-filter==24.3
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
httplib2==0.22.0
idna==3.10
jaconv==0.4.0
Jinja2==3.1.6
joblib==1.4.2
lxml==5.3.1
MarkupSafe==3.0.2
numpy==1.26.4
pandas==2.2.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pykakasi==2.3.0
pyparsing==3.2.1
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
soupsieve==2.6
sqlparse==0.5.3
stripe==11.6.0
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
wrapt==1.17.2
celery==5.3.6
mysqlclient==2.2.6
gunicorn==21.2.0
uvicorn==0.27.1
firebase-admin==6.6.0
langdetect==1.0.9
geoip2==4.8.0
django-redis==5.4.0
pymysql==1.1.1
weasyprint==62.3
reportlab==4.2.5
boto3==1.35.46
